Start Time

+++ Clash Royal Chest System/Assets/Scripts/Chest/MVC/ChestModel.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public class ChestModel
 {
     public Rewards Reward { get; private set; }
-    public float TimeToOpenInMin { get; private set; }
+    public int TimeToOpenInMin { get; private set; }
     public Sprite ChestClose { get; private set; }
     public Sprite ChestOpen { get; private set; }
